update to newest nightly, rebuilt as a lib
