+++ Cargo.toml
@@ -3,3 +3,6 @@
 name = "rust-sexp"
 version = "0.0.1"
 authors = ["Chris Gill <chrisfgill@gmail.com>", "Moritz Kaufmann"]
+
+[dependencies]
+regex = "*"
